+++ src/components/shared/routers/GameRouter.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { Redirect, Route } from "react-router-dom";
-import Game from "../../game/Game";
+import Lobby from "../../game/Lobby";
 
 const Container = styled.div`
   display: flex;
@@ -17,14 +17,14 @@ class GameRouter extends React.Component {
       <Container>
         <Route
           exact
-          path={`${this.props.base}/dashboard`}
-          render={() => <Game />}
+          path={`${this.props.base}/lobby`}
+          render={() => <Lobby />}
         />
 
         <Route
           exact
           path={`${this.props.base}`}
-          render={() => <Redirect to={`${this.props.base}/dashboard`} />}
+          render={() => <Redirect to={`${this.props.base}/lobby`} />}
         />
       </Container>
     );
